feat: Modify BurgerOrderScreen to use whopperPanel and togglePanels
