+++ BurgerKingKiosk/src/gui/UserStartManager.java
@@ -8,6 +8,9 @@ import javax.swing.JFrame;
 import javax.swing.JPanel;
 import javax.swing.SwingConstants;
 import javax.swing.border.EmptyBorder;
+
+import gui.burgerOrderScreen.BurgerOrderScreen;
+
 import java.awt.Color;
 import javax.swing.JLabel;
 import java.awt.event.MouseAdapter;
